# specs/002-fullstack-todo-app/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Full-Stack Todo Web Application API
  description: REST API for the multi-user todo application
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /api/auth/signup:
    post:
      summary: Create a new user account
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "secure_password"
              required:
                - email
                - password
      responses:
        '200':
          description: User account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    example: "123e4567-e89b-12d3-a456-426614174000"
                  email:
                    type: string
                    format: email
                    example: "user@example.com"
                  message:
                    type: string
                    example: "Account created successfully"
        '400':
          description: Invalid input or email already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Email already exists"

  /api/auth/signin:
    post:
      summary: Authenticate user and return token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "secure_password"
              required:
                - email
                - password
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    example: "123e4567-e89b-12d3-a456-426614174000"
                  email:
                    type: string
                    format: email
                    example: "user@example.com"
                  token:
                    type: string
                    example: "session_token"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid credentials"

  /api/todos:
    get:
      summary: Get all todos for the authenticated user
      tags:
        - Todos
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of todos for the user
          content:
            application/json:
              schema:
                type: object
                properties:
                  todos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Todo'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
    post:
      summary: Create a new todo
      tags:
        - Todos
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '200':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Title is required"
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"

  /api/todos/{id}:
    put:
      summary: Update an existing todo
      tags:
        - Todos
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Title is required"
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Todo not found"
    delete:
      summary: Delete a todo
      tags:
        - Todos
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Todo deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Todo deleted successfully"
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Todo not found"

  /api/todos/{id}/complete:
    patch:
      summary: Mark a todo as complete/incomplete
      tags:
        - Todos
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_completed:
                  type: boolean
                  example: true
              required:
                - is_completed
      responses:
        '200':
          description: Todo completion status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid completion status"
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Todo not found"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Todo:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          example: "Buy groceries"
        description:
          type: string
          example: "Milk, eggs, bread"
        is_completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2026-01-04T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-01-04T10:00:00Z"
        user_id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174001"
      required:
        - id
        - title
        - is_completed
        - created_at
        - updated_at
        - user_id

    CreateTodoRequest:
      type: object
      properties:
        title:
          type: string
          example: "New task"
        description:
          type: string
          example: "Optional description"
      required:
        - title

    UpdateTodoRequest:
      type: object
      properties:
        title:
          type: string
          example: "Updated task"
        description:
          type: string
          example: "Updated description"
        is_completed:
          type: boolean
          example: true